smart_energy_meter/ml/requirements.txt
pandas
scikit-learn
joblib
paho-mqtt
